+++ Implementation Phase/GELF/src/main/java/gelf/view/diagrams/components/SolidLine.java
@@ -3,7 +3,7 @@ package gelf.view.diagrams.components;
 import java.awt.Color;
 
 public class SolidLine extends DiagramLine {
-	protected SolidLine(PositionInFrame start, PositionInFrame end, Color color, double thickness) {
+	protected SolidLine(PositionInFrame start, PositionInFrame end, Color color, float thickness) {
 		super(start, end, color, thickness);
 	}
